# integration_package/docs/guides/openapi-v72.yaml
openapi: 3.1.0
info:
  title: CHE·NU™ V72 API
  description: |
    Governed Intelligence Operating System API
    
    ## GOUVERNANCE > EXÉCUTION
    
    All sensitive actions require checkpoint approval.
    
    ### Principles
    - No autonomous agents
    - Human-in-the-loop for critical actions
    - Complete audit trail
    - READ-ONLY XR interface
  version: '72.0.0'
  contact:
    name: CHE·NU Support
    email: api@chenu.io
  license:
    name: Proprietary
    url: https://chenu.io/license
servers:
  - url: http://localhost:8000/api/v1
    description: Development
  - url: https://staging.chenu.io/api/v1
    description: Staging
  - url: https://api.chenu.io/v1
    description: Production
...
security:
  - bearerAuth: []
paths:
...
  /threads:
    get:
      tags: [Threads]
      summary: List threads
      parameters:
        - name: sphere_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ThreadStatus'
        - name: maturity_level
          in: query
          schema:
            $ref: '#/components/schemas/MaturityLevel'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: List of threads
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreadsResponse'

    post:
      tags: [Threads]
      summary: Create thread
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateThreadRequest'
      responses:
        '201':
          description: Thread created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreadResponse'
  /threads/{threadId}:
    get:
      tags: [Threads]
      summary: Get thread by ID
      parameters:
        - $ref: '#/components/parameters/threadId'
      responses:
        '200':
          description: Thread details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreadResponse'
  /threads/{threadId}/archive:
    post:
      tags: [Threads]
      summary: Archive thread (requires checkpoint)
      parameters:
        - $ref: '#/components/parameters/threadId'
      responses:
        '200':
          description: Checkpoint created for approval
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckpointResponse'

  # ═══════════════════════════════════════════════════════════════════════════
  # DECISIONS
  # ═══════════════════════════════════════════════════════════════════════════
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    threadId:
      name: threadId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
...
    ThreadStatus:
      type: string
      enum: [active, paused, completed, archived]
    MaturityLevel:
      type: string
      enum: [SEED, SPROUTING, GROWING, MATURE, RIPE]
...
    CheckpointStatus:
      type: string
      enum: [pending, approved, rejected, expired]
    RiskLevel:
      type: string
      enum: [low, medium, high, critical]
...
    Thread:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        founding_intent:
          type: string
        sphere_id:
          type: string
        status:
          $ref: '#/components/schemas/ThreadStatus'
        maturity_level:
          $ref: '#/components/schemas/MaturityLevel'
        maturity_score:
          type: integer
        created_at:
          type: string
          format: date-time
...
    Checkpoint:
      type: object
      properties:
        id:
          type: string
          format: uuid
        action_type:
          type: string
        description:
          type: string
        requested_by:
          type: string
        status:
          $ref: '#/components/schemas/CheckpointStatus'
        risk_level:
          $ref: '#/components/schemas/RiskLevel'
        created_at:
          type: string
          format: date-time

    # Request bodies
...
    CreateThreadRequest:
      type: object
      required: [title, founding_intent, sphere_id]
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 200
        founding_intent:
          type: string
          minLength: 10
          maxLength: 1000
        sphere_id:
          type: string
...
    ThreadsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Thread'
    ThreadResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Thread'
...
    CheckpointResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Checkpoint'
...